Moved stacktrace util to sechub-commons-core #2280
- changed unit tests to junit5
- adopted unit tests to work without spring
