+++ sechub-scan-product-checkmarx/src/main/java/com/mercedesbenz/sechub/domain/scan/product/checkmarx/CheckmarxResilienceConsultant.java
@@ -9,12 +9,12 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 import org.springframework.web.client.HttpClientErrorException;
 
+import com.mercedesbenz.sechub.commons.core.util.StacktraceUtil;
 import com.mercedesbenz.sechub.sharedkernel.MustBeDocumented;
 import com.mercedesbenz.sechub.sharedkernel.resilience.ResilienceConsultant;
 import com.mercedesbenz.sechub.sharedkernel.resilience.ResilienceContext;
 import com.mercedesbenz.sechub.sharedkernel.resilience.ResilienceProposal;
 import com.mercedesbenz.sechub.sharedkernel.resilience.SimpleRetryResilienceProposal;
-import com.mercedesbenz.sechub.sharedkernel.util.StacktraceUtil;
 
 @Component
 public class CheckmarxResilienceConsultant implements ResilienceConsultant {
