+++ sechub-shared-kernel/src/main/java/com/mercedesbenz/sechub/sharedkernel/resilience/ResilientActionExecutor.java
@@ -10,7 +10,7 @@ import java.util.Objects;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import com.mercedesbenz.sechub.sharedkernel.util.StacktraceUtil;
+import com.mercedesbenz.sechub.commons.core.util.StacktraceUtil;
 
 /**
  * This class is able to execute actions which shall be resilient by help of
